--- .rubberduck/embedding/.gitignore
@@ -1,1 +1,0 @@
-*.json
