--- .rubberduck/template/drunken-pirate.rdt.md
@@ -1,1 +1,0 @@
-../../template/fun/drunken-pirate.rdt.md
