+++ app/vscode/asset/walkthrough/rubberduck-templates.md
@@ -1,14 +1,14 @@
-# Rubberduck Templates
+# PearAI Templates
 
 What if you want to craft an AI Chat that knows _specifically_ about your conventions?
 How cool would it be to have the answers in your own language?
 
-You can craft your own conversation templates by adding `.rdt.md` files to the `.rubberduck/template` folder in your workspace. See the [Rubberduck Template docs](https://github.com/rubberduck-ai/rubberduck-vscode/blob/main/doc/rubberduck-templates.md) for more information.
+You can craft your own conversation templates by adding `.rdt.md` files to the `.rubberduck/template` folder in your workspace. See the [PearAI Template docs](https://github.com/rubberduck-ai/rubberduck-vscode/blob/main/doc/rubberduck-templates.md) for more information.
 
-To use custom conversations, run the "Rubberduck: Start Custom Chat… 💬" command.
+To use custom conversations, run the "PearAI: Start Custom Chat… 💬" command.
 
 Here is an example of a [drunken pirate describing your code](https://github.com/rubberduck-ai/rubberduck-vscode/blob/main/template/fun/drunken-pirate.rdt.md):
 
 ![Describe code as a drunken pirate](https://raw.githubusercontent.com/rubberduck-ai/rubberduck-vscode/main/app/vscode/asset/media/drunken-pirate.gif)
 
-[Learn how to craft your own Rubberduck Template](https://github.com/rubberduck-ai/rubberduck-vscode/blob/main/doc/rubberduck-templates.md)!
+[Learn how to craft your own PearAI Template](https://github.com/rubberduck-ai/rubberduck-vscode/blob/main/doc/rubberduck-templates.md)!
