--- lib/extension/src/conversation/template/RubberduckTemplateLoadResult.ts
@@ -1,14 +1,0 @@
-import * as vscode from "vscode";
-import { RubberduckTemplate } from "./RubberduckTemplate";
-
-export type RubberduckTemplateLoadResult =
-  | {
-      type: "success";
-      file: vscode.Uri;
-      template: RubberduckTemplate;
-    }
-  | {
-      type: "error";
-      file: vscode.Uri;
-      error: unknown;
-    };
